+++ designer/server/src/common/components/page-panel-left/macro.njk
@@ -1,3 +1,3 @@
-{% macro pagePanelLeft(params) %}
+{% macro appPagePanelLeft(params) %}
   {%- include "./template.njk" -%}
 {% endmacro %}
